% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_in_categories.R
\name{count_in_categories}
\alias{count_in_categories}
\title{Count in categories for specific player id}
\usage{
count_in_categories(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Count in categories for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
countMe <- count_in_categories(account_id = id_r3m1ck)
}
}
