% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_player_win_lose.R
\name{count_player_win_lose}
\alias{count_player_win_lose}
\title{Counting win or lose for specific player id}
\usage{
count_player_win_lose(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Counting win or lose for specific player id
}
\examples{
\dontrun{
#Count r3m1ck's win and lose summaries
id_r3m1ck = "135474549"
count_win_lose <- count_player_win_lose(account_id = id_r3m1ck)
}
}
