% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hero_benchmarks.R
\name{get_hero_benchmarks}
\alias{get_hero_benchmarks}
\title{Get Hero Benchmarks}
\usage{
get_hero_benchmarks(hero_id)
}
\arguments{
\item{hero_id}{Dota2 hero id. You can get the hero id using \code{\link{get_heroes}}.}
}
\value{
GetHeroBenchmark returns a list, which includes the details like gold per minutes, xp per minutes, kills per minutes etc.
}
\description{
Get benchmark of single hero
}
\examples{
\dontrun{
hero_id = "47"
viperBenchmarks <- get_hero_benchmarks(hero_id)
}
}
