% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hero_rankings.R
\name{get_hero_rankings}
\alias{get_hero_rankings}
\title{Get heroes rankings for specific player id}
\usage{
get_hero_rankings(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get heroes rankings for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
rankings <- get_hero_rankings(account_id = id_r3m1ck)
}
}
