% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hero_stats.R
\name{get_hero_stats}
\alias{get_hero_stats}
\title{Get hero stats}
\usage{
get_hero_stats(account_id)
}
\description{
Get hero stats
}
\examples{
id_r3m1ck = "135474549"
heroStats <- get_hero_rankings(account_id = id_r3m1ck)
}
