% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heroes.R
\name{get_heroes}
\alias{get_heroes}
\title{Get total status for specific player id}
\usage{
get_heroes()
}
\description{
Get total status for specific player id
}
\examples{
\dontrun{
heroes <- get_heroes()
}
}
