% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_items.R
\name{get_items}
\alias{get_items}
\title{Get Dota2 Item List}
\usage{
get_items()
}
\value{
get_items returns a list of in-game items, which include the columns of name, cost, secret_shop, side_shop, recipe and localized name.
}
\description{
Get a list of in-game items
}
\examples{
\dontrun{
itemList <- get_items()
head(itemList)
}
}
