% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_details.R
\name{get_match_details}
\alias{get_match_details}
\title{Get Match Details}
\usage{
get_match_details(match_id)
}
\arguments{
\item{match_id}{Dota2 match id. You can get them of a specific player using \code{\link{get_matches}}.}
}
\value{
GetMatchDetails returns a list, which includes the details like result, duration and etc.
}
\description{
Get details of a single match
}
\examples{
\dontrun{
match_id = "3114150257"
matchDetails <- get_match_details(match_id)
}
}
