% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_matches}
\alias{get_matches}
\title{Get Matches for specific player id}
\usage{
get_matches(account_id, limit)
}
\arguments{
\item{account_id}{Steam ID}

\item{limit}{matches limit}
}
\description{
Get Matches for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
limit = 5
matches <- get_matches(account_id = id_r3m1ck, limit = limit)
}
}
