% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_played_heroes.R
\name{get_played_heroes}
\alias{get_played_heroes}
\title{Get played heroes for specific player id}
\usage{
get_played_heroes(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get played heroes for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
playedHeroes <- get_played_heroes(account_id = id_r3m1ck)
}
}
