% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_played_with.R
\name{get_played_with}
\alias{get_played_with}
\title{Get played with for specific player id}
\usage{
get_played_with(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get played with for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
playedWith <- get_played_with(account_id = id_r3m1ck)
}
}
