% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_ratings.R
\name{get_player_ratings}
\alias{get_player_ratings}
\title{Get ratings for specific player id}
\usage{
get_player_ratings(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get ratings for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
ratings <- get_player_ratings(account_id = id_r3m1ck)
}
}
