% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_summaries.R
\name{get_player_summaries}
\alias{get_player_summaries}
\title{Get Summaries a Specific Player}
\usage{
get_player_summaries(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get Summaries a Specific Player
}
\examples{
\dontrun{
#Get r3m1ck's profile summaries
id_r3m1ck = "135474549"
playerSummaries <- get_player_summaries(account_id = id_r3m1ck)
#Print the in-game name of r3m1ck
playerSummaries$profile$personaname
}
}
