% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_recent_matches.R
\name{get_recent_matches}
\alias{get_recent_matches}
\title{Get recent matches for specific player id}
\usage{
get_recent_matches(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get recent matches for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
recentMatches <- get_recent_matches(account_id = id_r3m1ck)
}
}
