% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wardmap.R
\name{get_wardmap}
\alias{get_wardmap}
\title{Get ward map for specific player id}
\usage{
get_wardmap(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get ward map for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
wardMap <- get_wardmap(account_id = id_r3m1ck)
}
}
