% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_player_by_hero.R
\name{top_player_by_hero}
\alias{top_player_by_hero}
\title{Get heroes rankings for specific player id}
\usage{
top_player_by_hero(hero_id)
}
\arguments{
\item{hero_id}{Hero Id}
}
\description{
Get heroes rankings for specific player id
}
\examples{
\dontrun{
viperId = "47"
topPlayer <- top_player_by_hero(hero = viperId)
}
}
