% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_stats.R
\name{total_stats}
\alias{total_stats}
\title{Get total status for specific player id}
\usage{
total_stats(account_id)
}
\arguments{
\item{account_id}{Steam ID}
}
\description{
Get total status for specific player id
}
\examples{
\dontrun{
id_r3m1ck = "135474549"
totalStats <- total_stats(account_id = id_r3m1ck)
}
}
